app/services/image_service.py
from PIL import Image
import os
from flask import current_app
from werkzeug.utils import secure_filename

class ImageService:
    def __init__(self):
        self.upload_folder = current_app.config['UPLOAD_FOLDER']
        self.allowed_extensions = {'png', 'jpg', 'jpeg', 'gif'}
        self.max_size = (800, 800)
        self.thumbnail_size = (200, 200)

    def save_image(self, image_file, item_id):
        pass

    def create_thumbnail(self, image_path):
        pass

    def delete_image(self, image_path):
        pass

    def get_image_path(self, item_id, filename):
        pass

    def allowed_file(self, filename):
        pass

    def optimize_image(self, image):
        pass

    def validate_image(self, image_file):
        pass

    def process_image(self, image_file, item_id):
        pass

    def get_image_metadata(self, image_path):
        pass

    def resize_image(self, image, size):
        pass

    def convert_format(self, image, format='JPEG'):
        pass

    def cleanup_unused_images(self):
        pass

    def generate_filename(self, original_filename, item_id):
        pass
