requirements.txt
Flask==3.0.0
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
marshmallow==3.15.0
pillow==8.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.26.0
Werkzeug==3.0.1